Increasing minAllowedWeights from 600 to 1200
Param: minAllowedWeights
Initial Value: 600
Suggested Value: 1200

Background:
By the whitepaper definition, the network has reached consensus when the majority of incentives are accrued to peers with trust exceeding 50% of the stake weight. The whitepaper outlines in some detail how this point creates deflationary pressure on a disjoint cabal.  We compute the distance from this point (below) showing a result of 0.5 which suggests the network is strongly out of consensus. Note: negative values are desirable.

>>> 0.5 - meta.I[ meta.T > 0.5 ].sum() = 0.5

Motivation for suggested change:

A peer's naive trust score is calculated by summing the number of inward non-zero weights. As a result, the total number of non-zero edges in the network, a.k.a the density of the weight matrix, directly reflects the expected trust scores for peers. Specifically, increasing minAllowedWeights from 600 to 1024 will increase trust and bring the network closer to consensus.


Cu