Increase Sequence Length & Decrease BatchSize
The current setup favours small models due to the short sequence length, but also forces operators to use smaller models due to the high batch size. Longer sequence lengths means that bigger AI models have a better chance for doing one-shots, and since bigger AI models can handle a minimum of 1024 tokens, this should eventually be pushed to the maximum size (2048).
 
This PR als decreases the amount of requests being sent, since increasing batch size or sequence length increases the memory burden on miners. Decreasing batch size also decreases the repetition on models, making the power usage lower and also making it possible for bigger models to reply on smaller hardware.

This PR is not aimed at creating a bigger burden for GPU's, but rather aimed at making bigger models more favourable due to the longer input to be used for the one-shot.